+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/items/BeeStinger.java
@@ -1,18 +1,15 @@
 package com.telepathicgrunt.the_bumblezone.items;
 
 import com.telepathicgrunt.the_bumblezone.entities.nonliving.BeeStingerEntity;
-import com.telepathicgrunt.the_bumblezone.events.player.PlayerLocateProjectileEvent;
 import com.telepathicgrunt.the_bumblezone.modinit.BzItems;
 import com.telepathicgrunt.the_bumblezone.modules.EntityMiscHandler;
 import com.telepathicgrunt.the_bumblezone.platform.BzArrowItem;
 import com.telepathicgrunt.the_bumblezone.utils.OptionalBoolean;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.projectile.AbstractArrow;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Items;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.level.Level;
@@ -35,18 +32,5 @@ public class BeeStinger extends BzArrowItem {
     public OptionalBoolean bz$isInfinite(ItemStack stack, ItemStack bow, Player player) {
         int enchantLevel = EnchantmentHelper.getItemEnchantmentLevel(Enchantments.INFINITY_ARROWS, bow);
         return OptionalBoolean.of(enchantLevel > 0);
-    }
-
-    public static ItemStack bowUsable(ItemStack ignoredAmmo, PlayerLocateProjectileEvent event) {
-        if (event.shooter() instanceof Player player && (event.weapon().is(Items.BOW) || event.weapon().is(Items.CROSSBOW))) {
-            Inventory inventory = player.getInventory();
-            for(int i = 0; i < inventory.getContainerSize(); ++i) {
-                ItemStack itemstack1 = inventory.getItem(i);
-                if (itemstack1.is(BzItems.BEE_STINGER.get())) {
-                    return itemstack1;
-                }
-            }
-        }
-        return null;
     }
 }